% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmldeps_rmarkdown.R
\name{html_dependency_bootstrap}
\alias{html_dependency_bootstrap}
\title{An HTML dependency using `bootstrap.js`}
\usage{
html_dependency_bootstrap(theme)
}
\arguments{
\item{theme}{the name of the bootstrap theme.}
}
\description{
An HTML dependency using `bootstrap.js`
}
