% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmldeps_rmarkdown.R
\name{html_dependency_font_awesome}
\alias{html_dependency_font_awesome}
\title{An HTML dependency for FontAwesome}
\usage{
html_dependency_font_awesome()
}
\description{
An HTML dependency for FontAwesome
}
