% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmldeps_rmarkdown.R
\name{html_dependency_jqueryui}
\alias{html_dependency_jqueryui}
\title{An HTML dependency using `jqueryui.js`}
\usage{
html_dependency_jqueryui()
}
\description{
An HTML dependency using `jqueryui.js`
}
