% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmldeps_rmarkdown.R
\name{html_dependency_navigation}
\alias{html_dependency_navigation}
\title{An HTML dependency for navigation}
\usage{
html_dependency_navigation(code_menu, source_embed)
}
\description{
We can build a script list that contains
`tabsets.js` and either of (or both)
`codefolding.js` or `sourceembed.js`.
}
