% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmldeps-package.R
\docType{package}
\name{htmldeps-package}
\alias{htmldeps}
\alias{htmldeps-package}
\title{The htmldeps package provides centralized functions for
HTML dependencies for several packages}
\description{
The htmldeps package provides centralized functions for
HTML dependencies for several packages
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/htmldeps}
  \item Report bugs at \url{https://github.com/rstudio/htmldeps/issues}
}

}
\author{
\strong{Maintainer}: Richard Iannone \email{rich@rstudio.com} (0000-0003-3925-190X)

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
