% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmldeps-package.R
\docType{package}
\name{htmldeps-package}
\alias{htmldeps}
\alias{htmldeps-package}
\title{The htmldeps package provides centralized functions for
HTML dependencies for several packages}
\description{
The htmldeps package provides centralized functions for
HTML dependencies for several packages
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/htmldeps}
  \item Report bugs at \url{https://github.com/rstudio/htmldeps/issues}
}

}
\author{
\strong{Maintainer}: Richard Iannone \email{rich@rstudio.com} (0000-0003-3925-190X)

Other contributors:
\itemize{
  \item RStudio, Inc. [copyright holder, funder]
  \item  jQuery Foundation (jQuery library) [copyright holder]
  \item  jQuery contributors (jQuery library; authors listed in inst/rmd/h/jquery-AUTHORS.txt) [contributor, copyright holder]
  \item  jQuery UI contributors (jQuery UI library; authors listed in inst/rmd/h/jqueryui-AUTHORS.txt) [contributor, copyright holder]
  \item Mark Otto (Bootstrap library) [contributor]
  \item Jacob Thornton (Bootstrap library) [contributor]
  \item  Bootstrap contributors (Bootstrap library) [contributor]
  \item  Twitter, Inc (Bootstrap library) [copyright holder]
  \item Alexander Farkas (html5shiv library) [contributor, copyright holder]
  \item Scott Jehl (respond.js library) [contributor, copyright holder]
  \item Greg Franko (tocify library) [contributor, copyright holder]
  \item Dave Gandy (Font-Awesome) [contributor, copyright holder]
  \item Ben Sperry (Ionicons) [contributor]
  \item  Drifty (Ionicons) [copyright holder]
}

}
\keyword{internal}
